refactor: extract components, utilities and constants from AdmittedCaseV2
